moved lane separation logic into a dedicated util class. No modifications to the code has been made.
